Remove policyable protocol. Add policies. Remove all deprecated resolve functions
