+++ Sources/Core/Transaction/CodableTransaction.swift
@@ -7,6 +7,7 @@
 import Foundation
 import BigInt
 
+
 ///  Structure capable of carying the parameters for any transaction type.
 ///  while all fields in this struct are optional, they are not necessarily
 ///  optional for the type of transaction they apply to.
@@ -187,22 +188,6 @@ public struct CodableTransaction {
         return self.envelope.encode(for: type)
     }
 
-    @available(*, deprecated, message: "Please use PolicyResolver instead")
-    public mutating func resolve(provider: Web3Provider) async {
-        // FIXME: Delete force try
-        self.gasLimit = try! await self.gasLimitPolicy.resolve(provider: provider, transaction: self)
-
-        if from != nil || sender != nil {
-            self.nonce = try! await self.resolveNonce(provider: provider)
-        }
-        if case .eip1559 = type {
-            self.maxFeePerGas = try! await self.maxFeePerGasPolicy.resolve(provider: provider)
-            self.maxPriorityFeePerGas = try! await self.maxPriorityFeePerGasPolicy.resolve(provider: provider)
-        } else {
-            self.gasPrice = try! await self.gasPricePolicy.resolve(provider: provider)
-        }
-    }
-
     public var noncePolicy: NoncePolicy
     public var maxFeePerGasPolicy: FeePerGasPolicy
     public var maxPriorityFeePerGasPolicy: PriorityFeePerGasPolicy
@@ -264,7 +249,7 @@ extension CodableTransaction: Codable {
         if let accessList = accessList, !accessList.isEmpty {
             try containier.encode(accessList, forKey: .accessList)
         }
-
+        
         if !gasLimit.isZero {
             try containier.encode(gasLimit.hexString, forKey: .gasLimit)
         }
@@ -288,97 +273,6 @@ extension CodableTransaction: Codable {
 
         if let from = from {
             try containier.encode(from, forKey: .from)
-        }
-    }
-
-}
-
-extension CodableTransaction {
-    public enum GasLimitPolicy {
-        case automatic
-        case manual(BigUInt)
-        case limited(BigUInt)
-        case withMargin(Double)
-
-        @available(*, deprecated, message: "Please use PolicyResolver instead")
-        func resolve(provider: Web3Provider, transaction: CodableTransaction?) async throws -> BigUInt {
-            guard let transaction = transaction else { throw Web3Error.valueError }
-            let request: APIRequest = .estimateGas(transaction, transaction.callOnBlock ?? .latest)
-            let response: APIResponse<BigUInt> = try await APIRequest.sendRequest(with: provider, for: request)
-            switch self {
-            case .automatic, .withMargin:
-                return response.result
-            case .manual(let value):
-                return value
-            case .limited(let limit):
-                if limit <= response.result {
-                    return response.result
-                } else {
-                    return limit
-                }
-            }
-        }
-    }
-
-    public enum GasPricePolicy {
-        case automatic
-        case manual(BigUInt)
-        case withMargin(Double)
-
-        @available(*, deprecated, message: "Please use PolicyResolver instead")
-        func resolve(provider: Web3Provider, transaction: CodableTransaction? = nil) async throws -> BigUInt {
-            let oracle = Oracle(provider)
-            switch self {
-            case .automatic, .withMargin:
-                return await oracle.gasPriceLegacyPercentiles().max() ?? 0
-            case .manual(let value):
-                return value
-            }
-        }
-    }
-
-    public enum PriorityFeePerGasPolicy: Policyable {
-        case automatic
-        case manual(BigUInt)
-
-        @available(*, deprecated, message: "Please use PolicyResolver instead")
-        public func resolve(provider: Web3Provider, transaction: CodableTransaction? = nil) async throws -> BigUInt {
-            let oracle = Oracle(provider)
-            switch self {
-            case .automatic:
-                return await oracle.tipFeePercentiles().max() ?? 0
-            case .manual(let value):
-                return value
-            }
-        }
-    }
-
-    public enum FeePerGasPolicy: Policyable {
-        case automatic
-        case manual(BigUInt)
-
-        @available(*, deprecated, message: "Please use PolicyResolver instead")
-        public func resolve(provider: Web3Provider, transaction: CodableTransaction? = nil) async throws -> BigUInt {
-            let oracle = Oracle(provider)
-            switch self {
-            case .automatic:
-                return await oracle.baseFeePercentiles().max() ?? 0
-            case .manual(let value):
-                return value
-            }
-        }
-    }
-
-    @available(*, deprecated, message: "Please use PolicyResolver instead")
-    func resolveNonce(provider: Web3Provider) async throws -> BigUInt {
-        switch noncePolicy {
-        case .pending, .latest, .earliest:
-            guard let address = from ?? sender else { throw Web3Error.valueError }
-            let request: APIRequest = .getTransactionCount(address.address, callOnBlock ?? .latest)
-            let response: APIResponse<BigUInt> = try await APIRequest.sendRequest(with: provider, for: request)
-            return response.result
-        case .exact(let value):
-            return value
         }
     }
 
@@ -413,7 +307,7 @@ extension CodableTransaction {
     ///   - parameters: EthereumParameters object containing additional parametrs for the transaction like gas
     public init(type: TransactionType? = nil, to: EthereumAddress, nonce: BigUInt = 0,
                 chainID: BigUInt = 0, value: BigUInt = 0, data: Data = Data(),
-                gasLimit: BigUInt = 0, maxFeePerGas: BigUInt? = nil, maxPriorityFeePerGas: BigUInt? = nil, gasPrice: BigUInt? = nil,
+                gasLimit: BigUInt = 0, maxFeePerGas: BigUInt? = nil, maxPriorityFeePerGas: BigUInt? = nil,  gasPrice: BigUInt? = nil,
                 accessList: [AccessListEntry]? = nil, v: BigUInt = 1, r: BigUInt = 0, s: BigUInt = 0) {
         // FIXME: This is duplication and should be fixed.
         self.data = data
