+++ Sources/web3swift/Operations/ReadTransaction.swift
@@ -34,7 +34,9 @@ public class ReadOperation {
 
     // TODO: Remove type erasing here, some broad wide protocol should be added instead
     public func callContractMethod() async throws -> [String: Any] {
-        await transaction.resolve(provider: web3.provider)
+        // TODO: might change to be dependency
+        let resolver = PolicyResolver(provider: web3.provider)
+        try await resolver.resolveAll(for: &transaction)
         // MARK: Read data from ABI flow
         // FIXME: This should be dropped, and after `execute()` call, just to decode raw data.
         let data: Data = try await self.web3.eth.callTransaction(transaction)
